Replace InlineRecommendationBox with InlineArticleCallout component. Update index file to export the new component and its type
